Fix RSAPSS salt length check
This commit fixes the salt length check for RSAPSS. The previous logic
was only validating rsa.PSSSaltLengthAuto or the length of the digest,
e.g. 32 for SHA256. But rsa.PSSSaltLengthEqualsHash was invalid, and
this should be equivalent to the length of the digest.

This commit allows using the TLS client handshake with an RSA key
because the current logic in Go uses RSAPSS with RSA keys and sets the
SaltLenth to rsa.PSSSaltLengthEqualsHash by default. See
https://github.com/golang/go/blob/b702e0438ae9577dcc642ba7696a89799c86c8b7/src/crypto/tls/handshake_client_tls13.go#L658-L675