Use credential activation encryption block size patch
The `legacy/tpm2` package was using the length of the symmetric
key as the length for the IV. This is wrong, as AES256 has a 32
byte key, but the block size is 16.

By using `github.com/google/go-tpm@7d0adf0a5e3b`, we get the
correct IV size. The change was approved already, but it's waiting
to be merged.